Moved loggerFor and other useful Kotlin extensions into KotilnUtils.kt and moved LogHelper into test-utils
