+++ finance/src/test/kotlin/net/corda/contracts/asset/CashTests.kt
@@ -13,7 +13,6 @@ import net.corda.core.node.services.unconsumedStates
 import net.corda.core.utilities.OpaqueBytes
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.WireTransaction
-import net.corda.core.utilities.*
 import net.corda.node.services.vault.NodeVaultService
 import net.corda.node.utilities.configureDatabase
 import net.corda.node.utilities.transaction
