+++ node/src/test/kotlin/net/corda/node/services/transactions/DistributedImmutableMapTests.kt
@@ -8,9 +8,8 @@ import io.atomix.copycat.server.CopycatServer
 import io.atomix.copycat.server.storage.Storage
 import io.atomix.copycat.server.storage.StorageLevel
 import net.corda.core.getOrThrow
-import net.corda.core.utilities.LogHelper
+import net.corda.testing.LogHelper
 import net.corda.node.services.network.NetworkMapService
-import net.corda.node.services.transactions.DistributedImmutableMap
 import net.corda.node.utilities.configureDatabase
 import net.corda.testing.freeLocalHostAndPort
 import net.corda.testing.node.makeTestDataSourceProperties
