+++ node/src/test/kotlin/net/corda/node/services/vault/NodeVaultServiceTest.kt
@@ -11,7 +11,7 @@ import net.corda.core.node.services.unconsumedStates
 import net.corda.core.utilities.OpaqueBytes
 import net.corda.core.transactions.SignedTransaction
 import net.corda.testing.DUMMY_NOTARY
-import net.corda.core.utilities.LogHelper
+import net.corda.testing.LogHelper
 import net.corda.node.utilities.configureDatabase
 import net.corda.node.utilities.transaction
 import net.corda.testing.BOC
