+++ node/src/test/kotlin/net/corda/node/services/vault/VaultWithCashTest.kt
@@ -16,7 +16,7 @@ import net.corda.core.transactions.SignedTransaction
 import net.corda.testing.BOB
 import net.corda.testing.DUMMY_NOTARY
 import net.corda.testing.DUMMY_NOTARY_KEY
-import net.corda.core.utilities.LogHelper
+import net.corda.testing.LogHelper
 import net.corda.node.utilities.configureDatabase
 import net.corda.node.utilities.transaction
 import net.corda.testing.MEGA_CORP
